Add unix_timestamp to QR payload

+++ app/src/main/java/stockholm/makerspace/boxterminator/network/ValidateBoxRequest.kt
@@ -1,4 +1,4 @@
 package stockholm.makerspace.boxterminator.network
 
 data class ValidateBoxRequest(val member_number: Int,
-                              val box_label_id: Int)
+                              val box_label_id: Long)
